refactored project to components
