+++ index.html
@@ -9,5 +9,5 @@
 <body>
     <div id="root"></div>
 </body>
-<script type="module" src="/App.js"
+<script type="module" src="/src/App.js"></script>
 </html>
