+++ package.json
@@ -18,6 +18,7 @@
     "parcel": "^2.15.1"
   },
   "dependencies": {
+    "json-server": "^1.0.0-beta.3",
     "react": "^19.1.0",
     "react-dom": "^19.1.0"
   }
