+++ src/components/Body.js
@@ -1,0 +1,12 @@
+import SearchBar from './SearchBar'
+import RestaurantCardsContainer from './RestaurantCardsContainer'
+const Body = ()=>{
+    return (
+        <div className="body">
+            <SearchBar/>
+            <RestaurantCardsContainer/>
+        </div>
+    )
+}
+
+export default Body
