+++ src/components/Header.js
@@ -1,0 +1,12 @@
+import Logo from './Logo'
+import NavBar from './NavBar';
+const Header = ()=>{
+    return (
+       <div className="header">
+            <Logo/>
+            <NavBar/>
+       </div>
+    )
+}
+
+export default Header;
