+++ src/components/Logo.js
@@ -1,0 +1,8 @@
+const Logo = ()=>{
+    return (
+        <div className="logo-container">
+            <img className="logo" src="https://www.creativefabrica.com/wp-content/uploads/2018/10/Chef-restaurant-logo-by-DEEMKA-STUDIO-4-580x348.jpg"/>
+        </div>
+        )
+}
+export default Logo;
