+++ src/components/NavBar.js
@@ -1,0 +1,10 @@
+const NavBar = ()=>{
+    return ( <ul className="nav-items">
+                <li>Home</li>
+                <li>About</li>
+                <li>Contact Us</li>
+                <li>Cart</li>
+            </ul>)
+}
+
+export default NavBar;
